+++ packages/client/src/Components/Mapping/data/default-hex-style.js
